+++ l10n/de_DE.js
@@ -7,11 +7,17 @@ OC.L10N.register(
     "%s is the language code for the respective disclaimer translation":
     "%s ist der Kode der Sprache der entsprechenden Übersetzung",
 
+    "Add":
+    "Hinzufügen",
+
     "Agree":
     "Zustimmen",
 
     "Agree disclaimer":
     "Haftungsablehnung",
+
+    "Agreement text":
+    "Zustimmungstext",
 
     "Automatically expire cookies older than":
     "Automatisches Ablaufdatum für bestehende Cookies",
@@ -46,6 +52,9 @@ OC.L10N.register(
     "Default language for the text":
     "Standardsprache des Textes der Haftungsablehnung",
 
+    "Delete":
+    "Löschen",
+
     "Disclaimer of liability":
     "Haftungsablehnung",
 
@@ -64,20 +73,20 @@ OC.L10N.register(
     "General Terms and conditions":
     "Allgemeine Geschäftsbedingungen",
 
-    "I accept the %s1general terms and conditions%s2":
-    "Ich akzeptiere die %s1allgemeine Geschäftsbedingungen%s2",
+    "GTC":
+    "AGB",
 
-    "I have read and understood the %s1disclaimer of liability%s2":
-    "Ich akzeptiere die %s1Ablehnung der Haftung%s2",
+    "Here you can add your own agreement type by specifying the following properties":
+    "Hier können Sie Ihre eigene Vereinbarung hinzufügen. Folgende Eigenschaften müssen definiert werden",
 
-    "I have read and understood the %s1legal disclaimer%s2":
-    "Ich akzeptiere den %s1Rechtlichen Hinweis%s2",
+    "I accept the @s1general terms and conditions@s2":
+    "Ich akzeptiere die @s1allgemeine Geschäftsbedingungen@s2",
 
-    "If true, it indicates that the user is a returning visitor, who already accepted the disclaimer":
-    "Fals wahr, es heißt der Benutzer hat die Webseite bereits besucht und die Haftungsablehnung zugestimmt",
+    "I have read and understood the @s1disclaimer of liability@s2":
+    "Ich akzeptiere die @s1Ablehnung der Haftung@s2",
 
-    "It can be a two character code, ie: %s or a five character code, ie: %s":
-    "Es kann entweder einer zwei Zeichen-Kode, z.B.: %s oder einer fünf Zeichen-Kode, z.B.: %s sein",
+    "I have read and understood the @s1legal disclaimer@s2":
+    "Ich akzeptiere den @s1Rechtlichen Hinweis@s2",
 
     "If the file for the user language doesn't exist, then the file for the default language will be used":
     "Falls die Datei der Sprache des Benutzers nicht existieren wurde, dann wird die Datei der Standartsprache benutz",
@@ -85,11 +94,20 @@ OC.L10N.register(
     "If the text file is bigger than this size, then its contents won't appear completely in the dialog, so, adjust the size till it is adequate to your needs":
     "Falls die Text-Datei größerer als diesen Wert ist, dann werden ihre gesamte Ihnalte nicht auf den Dialog dargestellt, also, passen Sie ihn an bis Ihre Anforderungen erfühlt sind",
 
+    "If true, it indicates that the user is a returning visitor, who already accepted the disclaimer":
+    "Fals wahr, es heißt der Benutzer hat die Webseite bereits besucht und die Haftungsablehnung zugestimmt",
+
     "If you want to modify this limit, do it at your own risk by changing the value of the parameter: maxAppTxtFileSize inside the config.php file":
     "Falls Sie diese Beschränkung erhöhen möchten, ändern Sie, auf eigenes Risiko, den Wert  Parameters: maxAppTxtFileSize der Datei config.php",
 
     "In case that the file for the default language doesn't exist, then the user will see an error message when trying to read the text":
     "Sollte die Datei der Standardsprache nicht existieren, dann wird den Benutzer eine Fehlermeldung sehen, sobald er den Text versuch zu lesen",
+
+    "It can be a two character code, ie: %s or a five character code, ie: %s":
+    "Es kann entweder einer zwei Zeichen-Kode, z.B.: %s oder einer fünf Zeichen-Kode, z.B.: %s sein",
+
+    "It will be used for the title of text dialogs":
+    "Es wird der Titel des Textdialogs benutzt",
 
     "Legal disclaimer":
     "Rechtlicher Hinweis",
@@ -103,6 +121,9 @@ OC.L10N.register(
     "Maximum file size in megabytes for the text file":
     "Maximum Größe der Textdatei in Megabytes",
 
+    "Menu entry":
+    "Menütext",
+
     "Menu entry position":
     "Position innerhalb des Menüs",
 
@@ -112,8 +133,14 @@ OC.L10N.register(
     "Months":
     "Monate",
 
+    "Name":
+    "Name",
+
     "Neither the file: %s nor: %s exist":
     "Weder die Datei %s noch: %s wurden gefunden",
+
+    "New":
+    "Neu",
 
     "Next":
     "Vor",
@@ -121,14 +148,17 @@ OC.L10N.register(
     "None":
     "Nicht anzeigen",
 
+    "Note":
+    "Bemerkung",
+
     "Ok":
     "Ok",
 
     "On the contrary, if it is a returning user, who already accepted the disclaimer and there is an expiration time for old cookies, then the age of the AGChecked cookie will be determined by getting the AGLastVisit cookie; if it is higher than the expiration date for old cookies, then it will be invalidated":
     "Im Gegenteil, falls den Benutzer die Webseite bereits besucht und die Haftungsablehnung zugestimmt hat, dann wird das Alter des Cookies durch den Wert von AGLastVisit gerechnet. Falls es höher als das Ablaufdatum der bisherigen Cookies, dann wird es gelöscht",
 
-    "Please read and agree the disclaimer before proceeding":
-    "Bitte Haftungsablehnung vor der Anmeldung lesen und zustimmen",
+    "On the \"Selected\" column you can choose the agreement type to be used by your OwnCloud":
+    "Auf die Spalte: \"Selektiert\" können Sie die Vereinwarung wählen, die von OwnCloud benutzt wird",
 
     "Please also note that there is a hard coded limit of: %s Megabytes; this limit was set in order to prevent setting higher values that will make ownCloud to crash":
     "Bitte achten Sie, dass auch ein absolutes Limit von: %s Megabytes gibt. Diesen Wert wurde gesetzt um zu verhindern, dass höhere Werte ownCloud hängen",
@@ -136,8 +166,23 @@ OC.L10N.register(
     "Please contact the webmaster":
     "Bitte kontaktieren Sie den Website-Administrator",
 
+    "Please make sure that the placeholders: \"@s1\" and \"@s2\" are present in the \"Agreement text\"":
+    "Bitte stellen Sie sicher, dass \"Zustimmungstext\" die Trennzeichen: \"@s1\" und \"@s2\" enthaltet",
+
+    "Please note that the placeholders: @s1 and @s2 are necessary to indicate where the text link on the login page begins and ends respectively":
+    "Bitte merken Sie dass, die Trennzeichen: @s1 y @s2 benötigt werden um zu spezifizieren wo den Textlink der Anmeldung fängt und endet",
+
+    "Please read and agree the disclaimer before proceeding":
+    "Bitte Haftungsablehnung vor der Anmeldung lesen und zustimmen",
+
     "Prev":
     "Zurück",
+
+    "Selected":
+    "Selektiert",
+
+    "Short text, which will be used for the menu entry on the user pages":
+    "Kurztext für den Menü-Eintrag",
 
     "Show a link to a PDF file with the disclaimer":
     "Link zur PDF-Datei anzeigen",
@@ -148,8 +193,29 @@ OC.L10N.register(
     "Show/Hide help":
     "Hilfe anzeigen/ausblenden",
 
+    "Text that will appear on the login dialog":
+    "Text, dass auf die Anmeldungsseite erscheint",
+
+    "The default agreements: \"Disclaimer of liability\", \"Legal disclaimer\", and \"General Terms and conditions\" can't be deleted":
+    "Die standard Vereinbarungen: \"Haftungsablehnung\", \"Rechtlicher Hinweis\", \"Allgemeine Geschäftsbedingungen\" kann man nicht löschen",
+
+    "The fields: \"Name\" and \"Agreement text\" are obligatory":
+    "Die Felder: \"Name\" und \"Zustimmungstext\" sind obligatorisch",
+
+    "The placeholder: \"@s1\" must be before \"@s2\"":
+    "Das Trennzeichen: \"@s1\" muss vor \"@s2\" kommen",
+
+    "The placeholders: \"@s1\" and \"@s2\" can only appear once on the \"Agreement text\"":
+    "Die Trennzeichen: \"@s1\" und \"@s2\" dürfen nur einmal auf \"Zustimmungstext\" erscheinen",
+
+    "The texts of the columns: \"Name\", \"Menu entry\", and \"Agreement text\" must be given in English, then you need to add the respective translations to files under the \"l10n\" folder":
+    "Die Texte der Spalten: \"Name\", \"Menütext\" und \"Zustimmungstext\" muss man auf Englisch eingeben, dann muss man die entsprechende Übersetzungen auf die Dateien des Ordners \"l10n\" hinzufügen",
+
     "The value of the AGChecked will be set the first time that the user visits the website or when it has been expired":
     "Der Wert von AGChecked werde beim ersten Besuch des Benutzer oder wenn ihn abgealufen ist",
+
+    "This is the name of your own agreement":
+    "Name Ihrer Vereinbarung",
 
     "This must be set in order to prevent the download of big text files, which may block the login page":
     "Diesen Wert musst gesetzt werden um zu verhindern, dass das Hochladen von größere Text-Dateien die Anmeldung-Seite blockieren",
@@ -174,6 +240,9 @@ OC.L10N.register(
 
     "Where":
     "Wo",
+
+    "With the \"delete\" button you can remove your own agreements":
+    "Mit der Taste \"löschen\", können Sie Ihre eigene Vereinbeirung entfernen",
 
     "Wk":
     "KW",